fix point 5 move selectors to querySelectors.js file
