+++ src/controller/endGameConditions.js
@@ -1,11 +1,9 @@
 import { allBoatsCoordinates } from './battleshipGrid';
 import generateEndGameScore from './score';
 import { timer } from './infoPanel';
-
-const gameScreen = document.querySelector('.game-screen');
-const endGameScreen = document.querySelector('.end-game-screen');
-const endGameScreenHeader = document.querySelector('.end-game-header');
-const remainingAmmo = document.querySelector('.available-ammo');
+import {
+  gameScreen, endGameScreen, endGameScreenHeader, remainingAmmo,
+} from '../constants/querySelectors';
 
 let remainingBoatsCoordinates = [...allBoatsCoordinates];
