+++ src/controller/infoPanel.js
@@ -1,25 +1,12 @@
 /* eslint-disable import/no-mutable-exports */
 import generateEndGameScore from './score';
 import fireIcon from '../assets/icons/fireIcon.png';
-
-const gameScreen = document.querySelector('.game-screen');
-const endGameScreen = document.querySelector('.end-game-screen');
-const endGameScreenHeader = document.querySelector('.end-game-header');
-
-const ammoProgressBar = document.querySelector('.ammo-progress-bar-line');
-const countDownTimer = document.querySelector('.countdown-timer');
-const ammoInfo = document.querySelector('.available-ammo');
-
-const fourCellShip = document.querySelector('.four-cell-ship');
-const firstThreeCellShip = document.querySelector('.first-three-cells-ship');
-const secondThreeCellShip = document.querySelector('.second-three-cells-ship');
-const firstTwoCellShip = document.querySelector('.first-two-cells-ship');
-const secondTwoCellShip = document.querySelector('.second-two-cells-ship');
-const thirdTwoCellShip = document.querySelector('.third-two-cells-ship');
-const firstOneCellShip = document.querySelector('.first-one-cell-ship');
-const secondOneCellShip = document.querySelector('.second-one-cell-ship');
-const thirdOneCellShip = document.querySelector('.third-one-cell-ship');
-const fourthOneCellShip = document.querySelector('.fourth-one-cell-ship');
+import {
+  gameScreen, endGameScreen, endGameScreenHeader, ammoProgressBar, countDownTimer,
+  remainingAmmo as ammoInfo,
+  fourCellShip, firstThreeCellShip, secondThreeCellShip, firstTwoCellShip, secondTwoCellShip,
+  thirdTwoCellShip, firstOneCellShip, secondOneCellShip, thirdOneCellShip, fourthOneCellShip,
+} from '../constants/querySelectors';
 
 let availableAmmo = 30;
 let defaultAmmo = 0;
