+++ src/controller/userActionGrid.js
@@ -5,8 +5,7 @@ import {
   generateMissImage, generateMissImageAroundSunkBoat, generateHitImage,
   addSunkBoatImageToBattleship,
 } from './battleshipGrid';
-
-const playAgainButton = document.querySelector('.play-again-button');
+import { playAgainButton } from '../constants/querySelectors';
 
 function validateAttack(e) {
   for (let i = 0; i < boatsState.length; i++) {
